Remove the shoreside address from the settings page in favor of the list
The shoreside address is no longer in the settings page, as now there is a system for saving / recalling clients in general. However, if there is no shoreside client, it will try the default address for the MIT shoreside (from http://oceanai.mit.edu/kfisher/pmwiki/pmwiki.php?n=Site.PavLabNetwork).

If you'd like to change your shoreside address, click 'Forget' on the incorrect shoreside, add another client named exactly 'shoreside' with the correct address and port, and save that. From then on that client will be used as the main shoreside.