Add some functionality to the AppCasts page
The page itself isn't the best functionally, but eventually it will be. The backend however is a good start, with the new AppCast class able to parse AppCast strings into objects for displaying.

This commit also includes some cleanup in the string parsing code to make it more general, and moves the 'Map -> Array' function to the MoosMail provider for use in multiple pages.