Add an AppCast Details page
This update makes details of AppCasts viewable on their own page. All of the AppCastable apps on the master node will be discovered, and casts on additional nodes will be added as they are discovered (usually if they are viewed on pMarineViewer). Requests will be sent every 1.5 seconds for additional casts when the details page is open.

This setup assumes that APPCAST messages are bridged to the master node in the uFldShoreBroker config block.