Add an Auto Updater & MS AppCenter Analytics to the app
These are disableable in the Settings page for any reason. The default option is to auto update if available when next possible and report standard metrics.

The motivation behind this is because while the app works fine in a 'perfect' dev environment, there are bound to be bugs out on the field. Rather than manually updating every device by hand, the auto update should streamline the rollout of bug fixes and such.

This feature is still being refined and worked on, as the app seems to randomly choose when and when not to check and install updates.