Fix Auto Updates sometimes prematurely rolling back
After installing an update, the first thing the code expects to see is a confirmation that the install was successful or it will be rolled back. Because the sync() function is inside of a module (storage) that can sometimes take extra time to startup, CodePush thinks the install failed and rolls it back. Therefore, immediately after the application starts up tell CodePush that the app is ready.