Add some validation to the client address field
The regex (https://stackoverflow.com/a/23542162/1709894) is currently setup to match an IP address with a port. However, because of the numerous different ways to set an address (DNS, hosts, IPv4, IPv6, etc) we don't lock input to match.

We do warn the user if the name isn't a proper IP address to draw some attention, but most people adding clients with this app know what they are doing, so leave them be.